se realizan cambios por los siguientes motivos: se deja funcional el proceso de envio de mail para cuando se queira cambiar la contraseña, falta validar el numero de reintentos y que caduque el OTP

+++ COM.JOMA.EMP.QUERY/Dtos/MailRecuperarContrasenaQueryDto.cs
@@ -13,6 +13,6 @@ namespace COM.JOMA.EMP.QUERY.Dtos
         public string Cuerpo { set; get; }
         public string RucCompania { set; get; }
         public long IdEmpresa { set; get; }
-        public long TipoEnvioMail { set; get; }
+        public byte TipoEnvioMail { set; get; }
     }
 }
